requirements.txt
httpx[http2]>=0.24.0
requests>=2.28.0
rsa>=4.9
pycryptodome>=3.18.0
cryptography>=41.0.0
axolotl-curve25519>=0.4.1
